progress on k8s configuration
This adds the python script which generates the CloudFormation template
which is used to deploy and configure required infrastructure.

This will add the application manifest (deployment and service
configuration) for the stonks application. The userData.sh script will
need to be updated to pull the manifest from this repository once it is
pushed and available.